move routes to posts blueprint
